+++ ARAnalytics.h
@@ -89,33 +89,60 @@
 + (void)setupHeapAnalyticsWithApplicationID:(NSString *)appID;
 + (void)setupChartbeatWithApplicationID:(NSString *)appID;
 
+
 /// Add a provider manually
-+ (void)setupProvider:(ARAnalyticalProvider*)provider;
++ (void)setupProvider:(ARAnalyticalProvider *)provider;
+
+/// Remove a provider manually
++ (void)removeProvider:(ARAnalyticalProvider *)provider;
+
+/// Show all current providers
++ (NSSet *)currentProviders;
+
 
 /// Set a per user property
+/// @warning Deprecated, will be removed in next major release
 + (void)identifyUserwithID:(NSString *)userID andEmailAddress:(NSString *)email __attribute__((deprecated));
+
+/// Register a user and an associated email address, it is fine to send nils for either.
 + (void)identifyUserWithID:(NSString *)userID andEmailAddress:(NSString *)email;
 
+/// Set a per user property
 + (void)setUserProperty:(NSString *)property toValue:(NSString *)value;
-+ (void)incrementUserProperty:(NSString *)counterName byInt:(int)amount;
+
+/// Adds to a user property if support exists in the provider
++ (void)incrementUserProperty:(NSString *)counterName byInt:(NSInteger)amount;
 
 /// Submit user events to providers
 + (void)event:(NSString *)event;
+
+/// Submit user events to providers with additional properties
 + (void)event:(NSString *)event withProperties:(NSDictionary *)properties;
 
 /// Submit errors to providers
 + (void)error:(NSError *)error;
+
+/// Submit errors to providers with an associated message
 + (void)error:(NSError *)error withMessage:(NSString *)message;
 
 /// Monitor Navigation changes as page view
 + (void)pageView:(NSString *)pageTitle;
+
 #if TARGET_OS_IPHONE
-+ (void)monitorNavigationViewController:(UINavigationController *)controller;
+/// Monitor a navigation controller, submitting each [ARAnalytics pageView:] on didShowViewController
+/// @warning Deprecated in favour of monitorNavigationController:
++ (void)monitorNavigationViewController:(UINavigationController *)controller __attribute__((deprecated));
+
+/// Monitor a navigation controller, submitting each [ARAnalytics pageView:] on didShowViewController
++ (void)monitorNavigationController:(UINavigationController *)controller;
 #endif
 
 /// Let ARAnalytics deal with the timing of an event
 + (void)startTimingEvent:(NSString *)event;
+
+/// Trigger a finishing event for the timing
 + (void)finishTimingEvent:(NSString *)event;
+
 /// @warning the properites must not contain the key string `length` .
 + (void)finishTimingEvent:(NSString *)event withProperties:(NSDictionary *)properties;